% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nonlinICP.class.R
\name{summary.nonlinICP.class}
\alias{summary.nonlinICP.class}
\title{summary function}
\usage{
\method{summary}{nonlinICP.class}(object, ...)
}
\arguments{
\item{object}{object of class 'nonlinICP.class'.}

\item{...}{Additional inputs to generic summary function (not used).}
}
\description{
Summary functions for 'nonlinICP.class' objects.
}
\author{
Christina Heinze-Deml and Jonas Peters
}
